@@ -1,14 +1,14 @@
 import React, { useEffect } from "react";
 
-import { css } from '@emotion/react'
-import { backOrContinueStyle, errorBoxStyle, flexGapReplacementStyle } from '../cssStyles'
+import { css } from "@emotion/react";
+import { backOrContinueStyle, errorBoxStyle, flexGapReplacementStyle } from "../cssStyles";
 
-import { useDispatch, useSelector } from 'react-redux';
-import { selectWorkflows, setSelectedWorkflowIndex } from '../redux/videoSlice'
-import { selectFinishState, selectPageNumber } from '../redux/finishSlice'
+import { useAppDispatch, useAppSelector } from "../redux/store";
+import { selectWorkflows, setSelectedWorkflowIndex } from "../redux/videoSlice";
+import { selectFinishState, selectPageNumber } from "../redux/finishSlice";
 
-import { PageButton } from './Finish'
-import { faChevronLeft } from "@fortawesome/free-solid-svg-icons";
+import { PageButton } from "./Finish";
+import { LuChevronLeft } from "react-icons/lu";
 import { SaveAndProcessButton } from "./WorkflowConfiguration";
 import { selectStatus, selectError } from "../redux/workflowPostAndProcessSlice";
 import { selectStatus as saveSelectStatus, selectError as saveSelectError } from "../redux/workflowPostSlice";
@@ -16,103 +16,113 @@ import { httpRequestState, Workflow } from "../types";
 import { SaveButton } from "./Save";
 import { EmotionJSX } from "@emotion/react/types/jsx-namespace";
 
-import './../i18n/config';
-import { useTranslation } from 'react-i18next';
+import { useTranslation } from "react-i18next";
 import { Trans } from "react-i18next";
-import { FormControlLabel, Radio, RadioGroup, withStyles } from "@material-ui/core";
-import { selectTheme } from "../redux/themeSlice";
+import { FormControlLabel, Radio, RadioGroup } from "@mui/material";
+import { useTheme } from "../themes";
 
 /**
  * Allows the user to select a workflow
  */
-const WorkflowSelection : React.FC<{}> = () => {
+const WorkflowSelection: React.FC = () => {
 
   const { t } = useTranslation();
 
-  const dispatch = useDispatch();
+  const dispatch = useAppDispatch();
 
   // Initialite redux states
-  const workflows = useSelector(selectWorkflows)
-  const finishState = useSelector(selectFinishState)
-  const pageNumber = useSelector(selectPageNumber)
-  const theme = useSelector(selectTheme)
+  let workflows = useAppSelector(selectWorkflows);
+  // Need to make copy to handle undefined displayOrder values
+  workflows = [...workflows].sort((a, b) => {
+    return (b.displayOrder - a.displayOrder);
+  });
 
-  const postAndProcessWorkflowStatus = useSelector(selectStatus);
-  const postAndProcessError = useSelector(selectError)
-  const saveStatus = useSelector(saveSelectStatus);
-  const saveError = useSelector(saveSelectError)
+  const finishState = useAppSelector(selectFinishState);
+  const pageNumber = useAppSelector(selectPageNumber);
+  const theme = useTheme();
+
+  const postAndProcessWorkflowStatus = useAppSelector(selectStatus);
+  const postAndProcessError = useAppSelector(selectError);
+  const saveStatus = useAppSelector(saveSelectStatus);
+  const saveError = useAppSelector(saveSelectError);
 
   const workflowSelectionStyle = css({
-    padding: '20px',
-    display: (finishState === "Start processing" && pageNumber === 1) ? 'flex' : 'none',
-    flexDirection: 'column' as const,
-    justifyContent: 'center',
-    alignItems: 'center',
+    padding: "20px",
+    display: (finishState === "Start processing" && pageNumber === 1) ? "flex" : "none",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
     ...(flexGapReplacementStyle(30, false)),
-  })
+  });
 
   const workflowSelectionSelectionStyle = css({
-    display: 'flex',
-    flexDirection: 'column' as const,
-    alignItems: 'left',
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "left",
     ...(flexGapReplacementStyle(20, false)),
-    flexWrap: 'wrap',
-    maxHeight: '50vh',
-  })
+    flexWrap: "wrap",
+    maxHeight: "50vh",
+  });
 
   useEffect(() => {
     if (workflows.length >= 1) {
-      dispatch(setSelectedWorkflowIndex(workflows[0].id))
+      dispatch(setSelectedWorkflowIndex(workflows[0].id));
     }
-  }, [dispatch, workflows])
+  }, [dispatch, workflows]);
 
-  const handleWorkflowSelectChange = (event: { target: { value: string}; }) => {
-    dispatch(setSelectedWorkflowIndex(event.target.value))
+  const handleWorkflowSelectChange = (event: { target: { value: string; }; }) => {
+    dispatch(setSelectedWorkflowIndex(event.target.value));
   };
 
   // Layout template
-  const render = (topTitle: string, topText: {} | null | undefined, hasWorkflowButtons: boolean,
+  const render = (topTitle: string, topText: JSX.Element, hasWorkflowButtons: boolean,
     nextButton: EmotionJSX.Element, errorStatus: httpRequestState["status"],
     errorMessage: httpRequestState["error"]) => {
     return (
       <div css={workflowSelectionStyle}>
         <h2>{topTitle}</h2>
         {topText}
-        { hasWorkflowButtons &&
-            <RadioGroup
-              css={workflowSelectionSelectionStyle}
-              title="Workflow Selection Area"
-              defaultValue={workflows[0].id}
-              name="Workflow Selection Area"
-              onChange={handleWorkflowSelectChange}
-            >
-              {workflows.map( (workflow: Workflow, index: number) => (
-                <WorkflowButton key={workflow.id} stateName={workflow.name} workflowId={workflow.id} workflowDescription={workflow.description}/>
-              ))}
-            </RadioGroup>
+        {hasWorkflowButtons &&
+          <RadioGroup
+            css={workflowSelectionSelectionStyle}
+            defaultValue={workflows[0].id}
+            name="Workflow Selection Area"
+            onChange={handleWorkflowSelectChange}
+          >
+            {workflows.map((workflow: Workflow, _index: number) => (
+              <WorkflowButton
+                key={workflow.id}
+                stateName={workflow.name}
+                workflowId={workflow.id}
+                workflowDescription={workflow.description}
+              />
+            ))}
+          </RadioGroup>
         }
         <div css={backOrContinueStyle}>
-          <PageButton pageNumber={0} label={t("workflowSelection.back-button")} iconName={faChevronLeft}/>
+          <PageButton pageNumber={0} label={t("workflowSelection.back-button")} Icon={LuChevronLeft} />
           {/* <PageButton pageNumber={2} label="Continue" iconName={faChevronRight}/> */}
           {nextButton}
         </div>
         <div css={errorBoxStyle(errorStatus === "failed", theme)} role="alert">
           <span>{t("various.error-text")}</span><br />
-          {errorMessage ? t("various.error-details-text", {errorMessage: postAndProcessError}) : t("various.error-noDetails-text")}<br/>
+          {errorMessage ?
+            t("various.error-details-text", { errorMessage: postAndProcessError }) :
+            t("various.error-text")}<br />
         </div>
       </div>
     );
-  }
+  };
 
   // Fills the layout template with values based on how many workflows are available
   const renderSelection = () => {
     if (workflows.length <= 0) {
-      return(
+      return (
         render(
           t("workflowSelection.saveAndProcess-text"),
           <Trans i18nKey="workflowSelection.noWorkflows-text">
-            A problem occurred, there are no workflows to process your changes with.<br />
-            Please save your changes and contact an Opencast Administrator.
+            There are no workflows to process your changes with.<br />
+            Please save your changes and contact an administrator.
           </Trans>,
           false,
           <SaveButton />,
@@ -125,11 +135,11 @@ const WorkflowSelection : React.FC<{}> = () => {
         render(
           t("workflowSelection.saveAndProcess-text"),
           <Trans i18nKey="workflowSelection.oneWorkflow-text">
-            The video will be cut and processed with the workflow "{{workflow: workflows[0].name}}".<br/>
+            The video will be cut and processed with the workflow {{ workflow: workflows[0].name }}.<br />
             This will take some time.
           </Trans>,
           false,
-          <SaveAndProcessButton text={t("workflowSelection.startProcessing-button")}/>,
+          <SaveAndProcessButton text={t("workflowSelection.startProcessing-button")} />,
           postAndProcessWorkflowStatus,
           postAndProcessError
         )
@@ -142,33 +152,42 @@ const WorkflowSelection : React.FC<{}> = () => {
             {t("workflowSelection.manyWorkflows-text")}
           </div>,
           true,
-          <SaveAndProcessButton text= {t("workflowSelection.startProcessing-button")}/>,
+          <SaveAndProcessButton text={t("workflowSelection.startProcessing-button")} />,
           postAndProcessWorkflowStatus,
           postAndProcessError
         )
-      )
+      );
     }
-  }
+  };
 
   return (
     renderSelection()
   );
-}
+};
 
-const WorkflowButton: React.FC<{stateName: string, workflowId: string, workflowDescription: string}> = ({stateName, workflowId, workflowDescription}) => {
+const WorkflowButton: React.FC<{
+  stateName: string,
+  workflowId: string,
+  workflowDescription: string;
+}> = ({
+  stateName,
+  workflowId,
+  workflowDescription,
+}) => {
 
   // Note: Styling the Radio Button is done in WorkflowSelectRadio
 
   const labelStyle = css({
-    display: 'flex',
-    flexDirection: 'column',
-    maxWidth: '500px'
-  })
+    display: "flex",
+    flexDirection: "column",
+    maxWidth: "500px",
+    paddingTop: "2px",
+  });
 
   const headerStyle = css({
-    width: '100%',
-    padding: '5px',
-    fontSize: 'larger',
+    width: "100%",
+    padding: "5px 0px",
+    fontSize: "larger",
   });
 
   return (
@@ -181,17 +200,28 @@ const WorkflowButton: React.FC<{stateName: string, workflowId: string, workflowD
       }
     />
   );
-}
+};
 
-const WorkflowSelectRadio = withStyles({
-  root: {
-    alignSelf: 'start',
-    color: 'grey',
+const WorkflowSelectRadio: React.FC = props => {
+
+  const theme = useTheme();
+
+  const style = css({
+    alignSelf: "start",
+    color: `${theme.text}`,
     "&$checked": {
-      color: 'grey'
-    }
-  },
-  checked: {}
-})((props) => <Radio color="default" {...props} />);
+      color: `${theme.text}`,
+    },
+  });
+
+  return (
+    <Radio
+      color="default"
+      css={style}
+      {...props}
+    />
+  );
+};
+
 
 export default WorkflowSelection;
