@@ -2,15 +2,18 @@ import { PlaywrightTestConfig, devices } from '@playwright/test';
 
 const config: PlaywrightTestConfig = {
 
-  reporter: [ [process.env.CI ? 'github' : 'list',],
-            ['html', { outputFolder: 'playwright-report' }]],
+  reporter: [
+    [process.env.CI ? 'github' : 'list'],
+    ['html', { outputFolder: 'playwright-report' }],
+  ],
   testIgnore: '**/redux/**',
-  retries: 3,
+  retries: 1,
   timeout: 60 * 1000,
 
   use: {
     baseURL: 'http://localhost:3000/',
     headless: true,
+    screenshot: 'only-on-failure',
   },
 
   webServer: {
