@@ -1,15 +1,12 @@
-import React from 'react';
-import Body from './main/Body';
-import { GlobalStyle } from './cssStyles';
-import { useSelector } from 'react-redux';
-import { selectThemeState } from './redux/themeSlice';
+import Body from "./main/Body";
+import Header from "./main/Header";
+import { GlobalStyle } from "./cssStyles";
 
 function App() {
-  const theme = useSelector(selectThemeState);
-  
   return (
-    <div className="App" css={theme}>
+    <div className="App">
       <GlobalStyle />
+      <Header />
       <Body />
     </div>
   );
