@@ -4,22 +4,21 @@
  * - GET parameters
  * and exports them.
  * Code was largely adapted from https://github.com/elan-ev/opencast-studio/blob/master/src/settings.js (January 11th, 2021)
- * 
+ *
  * Also does some global hotkey configuration
  */
-import parseToml from '@iarna/toml/parse-string';
-import deepmerge from 'deepmerge';
-import { configure } from 'react-hotkeys';
-import { Flavor } from './types';
+import parseToml from "@iarna/toml/parse-string";
+import deepmerge from "deepmerge";
+import { Flavor } from "./types";
 
 /**
  * Local constants
  */
-const CONTEXT_SETTINGS_FILE = 'editor-settings.toml';
+const CONTEXT_SETTINGS_FILE = "editor-settings.toml";
 
 // Sources that values can come from.
-const SRC_SERVER = 'src-server';
-const SRC_URL = 'src-url';
+const SRC_SERVER = "src-server";
+const SRC_URL = "src-url";
 
 /**
  * Possible configuration values for a metadata catalog field
@@ -29,20 +28,31 @@ export interface configureFieldsAttributes {
   readonly: boolean,
 }
 
+export interface subtitleTags {
+  lang: string,
+  "auto-generated": string,
+  "auto-generator": string,
+  type: string,
+}
+
 /**
  * Settings interface
  */
 interface iSettings {
   id: string | undefined,
+  allowedCallbackPrefixes: string[],
+  callbackUrl: string | undefined,
+  callbackSystem: string | undefined,
   opencast: {
     url: string,
     name: string | undefined,
     password: string | undefined,
     audioFileFlavor: Flavor | undefined,
+    local: boolean,
   },
   metadata: {
     show: boolean,
-    configureFields: { [key: string]: { [key: string]: configureFieldsAttributes } }  | undefined,
+    configureFields: { [key: string]: { [key: string]: configureFieldsAttributes; }; } | undefined,
   },
   trackSelection: {
     show: boolean,
@@ -54,9 +64,10 @@ interface iSettings {
   subtitles: {
     show: boolean,
     mainFlavor: string,
-    languages: { [key: string]: string } | undefined,
+    languages: { [key: string]: subtitleTags; } | undefined,
+    icons: { [key: string]: string; } | undefined,
     defaultVideoFlavor: Flavor | undefined,
-  }
+  };
 }
 
 /**
@@ -66,13 +77,17 @@ interface iSettings {
  * urlParameterSettings: contains values from GET parameters
  * settings: contains the combined values from all other setting objects
  */
-var defaultSettings: iSettings = {
+const defaultSettings: iSettings = {
   id: undefined,
+  allowedCallbackPrefixes: [],
+  callbackUrl: undefined,
+  callbackSystem: undefined,
   opencast: {
     url: window.location.origin,
     name: undefined,
     password: undefined,
     audioFileFlavor: undefined,
+    local: true,
   },
   metadata: {
     show: true,
@@ -88,13 +103,14 @@ var defaultSettings: iSettings = {
   subtitles: {
     show: false,
     mainFlavor: "captions",
-    languages: undefined,
+    languages: {},
+    icons: undefined,
     defaultVideoFlavor: undefined,
-  }
-}
-var configFileSettings: iSettings
-var urlParameterSettings: iSettings
-export var settings: iSettings
+  },
+};
+let configFileSettings: iSettings;
+let urlParameterSettings: iSettings;
+export let settings: iSettings;
 
 /**
  * Entry point. Loads values from settings into the exported variables
@@ -104,30 +120,40 @@ export var settings: iSettings
  * 3. Default values
  */
 export const init = async () => {
+
+  // Get color scheme from local storage, otherwise set auto scheme based on preference
+  let scheme = window.localStorage.getItem("colorScheme");
+  if (scheme === null || !["light", "dark", "light-high-contrast", "dark-high-contrast"].includes(scheme)) {
+    const lightness = window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light";
+    const contrast = window.matchMedia("(prefers-contrast: more)").matches ? "-high-contrast" : "";
+    scheme = `${lightness}${contrast}`;
+  }
+  document.documentElement.dataset.colorScheme = scheme;
+
   // Get settings from config file
-  await loadContextSettings().then((result) => {
-    configFileSettings = validate(result, false, SRC_SERVER, "from server settings file")
-  })
+  await loadContextSettings().then(result => {
+    configFileSettings = validate(result, false, SRC_SERVER, "from server settings file");
+  });
 
   // Get settings from URL query.
-  var urlParams = new URLSearchParams(window.location.search);
+  const urlParams = new URLSearchParams(window.location.search);
 
-  let rawUrlSettings = {};
+  const rawUrlSettings = {};
   urlParams.forEach((value, key) => {
-    // Create empty objects for full path (if the key contains '.') and set
+    // Create empty objects for full path (if the key contains ".") and set
     // the value at the end.
-    let obj : {[k: string]: any} = rawUrlSettings;
-    if (key.startsWith('opencast.')) {
+    let obj: { [k: string]: any; } = rawUrlSettings;
+    if (key.startsWith("opencast.") || key === "allowedCallbackPrefixes") {
       return;
     }
 
     // Fallback for old parameter
-    if (key === 'mediaPackageId') {
-      key = 'id';
+    if (key === "mediaPackageId") {
+      key = "id";
     }
 
-    const segments = key.split('.');
-    segments.slice(0, -1).forEach((segment) => {
+    const segments = key.split(".");
+    segments.slice(0, -1).forEach(segment => {
       if (!(segment in obj)) {
         obj[segment] = {};
       }
@@ -136,29 +162,18 @@ export const init = async () => {
     obj[segments[segments.length - 1]] = value;
   });
 
-  urlParameterSettings = validate(rawUrlSettings, true, SRC_URL, 'given as URL GET parameter');
+  urlParameterSettings = validate(rawUrlSettings, true, SRC_URL, "given as URL GET parameter");
 
   // Combine results
   settings = merge.all([defaultSettings, configFileSettings, urlParameterSettings]) as iSettings;
 
-  // Configure hotkeys
-  configure({
-    ignoreTags: [],   // Do not ignore hotkeys when focused on a textarea, input, select
-    ignoreEventsCondition: (e: any) => {
-      // Ignore hotkeys when focused on a textarea, input, select IF that hotkey is expected to perform
-      // a certain function in that element that is more important than any hotkey function
-      // (e.g. you need "Space" in a textarea to create whitespaces, not play/pause videos)
-      if (e.target && e.target.tagName) {
-        const tagname = e.target.tagName.toLowerCase()
-        if ((tagname === "textarea" || tagname === "input" || tagname === "select")
-          && (!e.altKey && !e.ctrlKey)
-          && (e.code === "Space" || e.code === "ArrowLeft" || e.code === "ArrowRight" || e.code === "ArrowUp" || e.code === "ArrowDown")) {
-          return true
-        }
-      }
-      return false
-    },
-  })
+  // Prepare local setting to avoid complicated checks later
+  settings.opencast.local = settings.opencast.local && settings.opencast.url === window.location.origin;
+
+  // Prevent malicious callback urls
+  settings.callbackUrl = settings.allowedCallbackPrefixes.some(
+    p => settings.callbackUrl?.startsWith(p)
+  ) ? settings.callbackUrl : undefined;
 };
 
 /**
@@ -167,47 +182,47 @@ export const init = async () => {
 const loadContextSettings = async () => {
 
   // Try to retrieve the context settings.
-  let basepath = process.env.PUBLIC_URL || '/';
-  if (!basepath.endsWith('/')) {
-    basepath += '/';
+  let basepath = process.env.PUBLIC_URL || "/";
+  if (!basepath.endsWith("/")) {
+    basepath += "/";
   }
 
   // Construct path to settings file. If the `REACT_APP_SETTINGS_PATH` is
-  // given and starts with '/', it is interpreted as absolute path from the
+  // given and starts with "/", it is interpreted as absolute path from the
   // server root.
   const settingsPath = process.env.REACT_APP_SETTINGS_PATH || CONTEXT_SETTINGS_FILE;
-  const base = settingsPath.startsWith('/') ? '' : basepath;
+  const base = settingsPath.startsWith("/") ? "" : basepath;
   const url = `${window.location.origin}${base}${settingsPath}`;
   let response;
   try {
     response = await fetch(url);
   } catch (e) {
-    console.warn(`Could not access '${settingsPath}' due to network error!`, e || "");
+    console.warn(`Could not access "${settingsPath}" due to network error!`, e || "");
     return null;
   }
 
   if (response.status === 404) {
     // If the settings file was not found, we silently ignore the error. We
     // expect many installation to provide this file.
-    console.debug(`'${settingsPath}' returned 404: ignoring`);
+    console.debug(`"${settingsPath}" returned 404: ignoring`);
     return null;
   } else if (!response.ok) {
     console.error(
-      `Fetching '${settingsPath}' failed: ${response.status} ${response.statusText}`
+      `Fetching "${settingsPath}" failed: ${response.status} ${response.statusText}`
     );
     return null;
   }
 
-  if (response.headers.get('Content-Type')?.startsWith('text/html')) {
-    console.warn(`'${settingsPath}' request has 'Content-Type: text/html' -> ignoring...`);
+  if (response.headers.get("Content-Type")?.startsWith("text/html")) {
+    console.warn(`"${settingsPath}" request has "Content-Type: text/html" -> ignoring...`);
     return null;
   }
 
   try {
     return parseToml(await response.text());
   } catch (e) {
-    console.error(`Could not parse '${settingsPath}' as TOML: `, e);
-    throw new SyntaxError(`Could not parse '${settingsPath}' as TOML: ${e}`);
+    console.error(`Could not parse "${settingsPath}" as TOML: `, e);
+    throw new SyntaxError(`Could not parse "${settingsPath}" as TOML: ${e}`);
   }
 
 };
@@ -223,7 +238,7 @@ const validate = (obj: Record<string, any> | null, allowParse: boolean, src: str
   // Validates `obj` with `schema`. `path` is the current path used for error
   // messages.
   const validate = (schema: any, obj: Record<string, any> | null, path: string) => {
-    if (typeof schema === 'function') {
+    if (typeof schema === "function") {
       return validateValue(schema, obj, path);
     } else {
       return validateObj(schema, obj, path);
@@ -232,14 +247,18 @@ const validate = (obj: Record<string, any> | null, allowParse: boolean, src: str
 
   // Validate a settings value with a validation function. Returns the final
   // value of the setting or `null` if it should be ignored.
-  const validateValue = (validation: (arg0: any, arg1: boolean, arg2: string) => any, value: Record<string, any> | null, path: string) => {
+  const validateValue = (
+    validation: (arg0: any, arg1: boolean, arg2: string) => any,
+    value: Record<string, any> | null,
+    path: string
+  ) => {
     try {
       const newValue = validation(value, allowParse, src);
       return newValue === undefined ? value : newValue;
     } catch (e) {
       console.warn(
-        `Validation of setting '${path}' (${sourceDescription}) with value '${value}' failed: `
-          + `${e}. Ignoring.`
+        `Validation of setting "${path}" (${sourceDescription}) with value "${value}" failed: `
+        + `${e}. Ignoring.`
       );
       return null;
     }
@@ -250,7 +269,7 @@ const validate = (obj: Record<string, any> | null, allowParse: boolean, src: str
   const validateObj = (schema: any, obj: Record<string, any> | null, path: string) => {
     // We iterate through all keys of the given settings object, checking if
     // each key is valid and recursively validating the value of that key.
-    let out : {[k: string]: any} = {};
+    const out: { [k: string]: any; } = {};
     for (const key in obj) {
       const newPath = path ? `${path}.${key}` : key;
       if (key in schema) {
@@ -263,7 +282,7 @@ const validate = (obj: Record<string, any> | null, allowParse: boolean, src: str
         }
       } else {
         console.warn(
-          `'${newPath}' (${sourceDescription}) is not a valid settings key. Ignoring.`
+          `"${newPath}" (${sourceDescription}) is not a valid settings key. Ignoring.`
         );
       }
     }
@@ -272,66 +291,76 @@ const validate = (obj: Record<string, any> | null, allowParse: boolean, src: str
   };
 
   return validate(SCHEMA, obj, "");
-}
+};
 
 
 // Validation functions for different types.
 const types = {
-  'string': (v: any, allowParse: any) => {
-    if (typeof v !== 'string') {
+  "string": (v: any, _allowParse: any) => {
+    if (typeof v !== "string") {
       throw new Error("is not a string, but should be");
     }
   },
-  'boolean': (v: string, allowParse: any) => {
-    if (typeof v === 'boolean') {
+  "boolean": (v: string, allowParse: any) => {
+    if (typeof v === "boolean") {
       return;
     }
 
     if (allowParse) {
-      if (v === 'true') {
+      if (v === "true") {
         return true;
       }
-      if (v === 'false') {
+      if (v === "false") {
         return false;
       }
-      throw new Error("can't be parsed as boolean");
+      throw new Error("cant be parsed as boolean");
     } else {
       throw new Error("is not a boolean");
     }
   },
-  'map': (v: any, allowParse: any) => {
-    for (let key in v) {
-      if (typeof key !== 'string') {
-        throw new Error("is not a string, but should be");
-      }
-      if (typeof v[key] !== 'string') {
+  "array": (v: any, _allowParse: any) => {
+    if (!Array.isArray(v)) {
+      throw new Error("is not an array, but should be");
+    }
+    for (const entry in v) {
+      if (typeof entry !== "string") {
         throw new Error("is not a string, but should be");
       }
     }
   },
-  'objectsWithinObjects': (v: any, allowParse: any) => {
-    for (let catalogName in v) {
-      if (typeof catalogName !== 'string') {
+  "map": (v: any, _allowParse: any) => {
+    for (const key in v) {
+      if (typeof key !== "string") {
         throw new Error("is not a string, but should be");
       }
-      for (let fieldName in v[catalogName]) {
-        if (typeof fieldName !== 'string') {
+      if (typeof v[key] !== "string") {
+        throw new Error("is not a string, but should be");
+      }
+    }
+  },
+  "objectsWithinObjects": (v: any, _allowParse: any) => {
+    for (const catalogName in v) {
+      if (typeof catalogName !== "string") {
+        throw new Error("is not a string, but should be");
+      }
+      for (const fieldName in v[catalogName]) {
+        if (typeof fieldName !== "string") {
           throw new Error("is not a string, but should be");
         }
-        for (let attributeName in v[catalogName][fieldName]) {
-          if (typeof attributeName !== 'string') {
+        for (const attributeName in v[catalogName][fieldName]) {
+          if (typeof attributeName !== "string") {
             throw new Error("is not a string, but should be");
           }
-          if (attributeName === 'show' && typeof v[catalogName][fieldName][attributeName] !== 'boolean') {
+          if (attributeName === "show" && typeof v[catalogName][fieldName][attributeName] !== "boolean") {
             throw new Error("is not a boolean");
           }
-          if (attributeName === 'readonly' && typeof v[catalogName][fieldName][attributeName] !== 'boolean') {
+          if (attributeName === "readonly" && typeof v[catalogName][fieldName][attributeName] !== "boolean") {
             throw new Error("is not a boolean");
           }
         }
       }
     }
-  }
+  },
 };
 
 // Defines all potential settings and their types.
@@ -346,6 +375,9 @@ const types = {
 // above for some examples.
 const SCHEMA = {
   id: types.string,
+  allowedCallbackPrefixes: types.array,
+  callbackUrl: types.string,
+  callbackSystem: types.string,
   opencast: {
     url: types.string,
     name: types.string,
@@ -353,26 +385,27 @@ const SCHEMA = {
     audioFileFlavor: types.map,
   },
   metadata: {
-    show : types.boolean,
+    show: types.boolean,
     configureFields: types.objectsWithinObjects,
   },
   trackSelection: {
-    show : types.boolean,
+    show: types.boolean,
   },
   subtitles: {
     show: types.boolean,
     mainFlavor: types.string,
-    languages: types.map,
+    languages: types.objectsWithinObjects,
+    icons: types.map,
     defaultVideoFlavor: types.map,
   },
   thumbnail: {
-    show : types.boolean,
+    show: types.boolean,
     simpleMode: types.boolean,
-  }
-}
+  },
+};
 
 const merge = (a: iSettings, b: iSettings) => {
   return deepmerge(a, b, { arrayMerge });
 };
-merge.all = (array: object[]) => deepmerge.all(array, { arrayMerge })
-const arrayMerge = (destinationArray: any, sourceArray: any, options: any) => sourceArray;
+merge.all = (array: object[]) => deepmerge.all(array, { arrayMerge });
+const arrayMerge = (_destinationArray: any, sourceArray: any, _options: any) => sourceArray;
