@@ -1,39 +1,35 @@
 import React from "react";
 
-import { css } from '@emotion/react'
-import { basicButtonStyle, backOrContinueStyle, navigationButtonStyle, flexGapReplacementStyle} from '../cssStyles'
+import { css } from "@emotion/react";
+import { basicButtonStyle, backOrContinueStyle, navigationButtonStyle, flexGapReplacementStyle } from "../cssStyles";
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faChevronLeft, faTimesCircle
-} from "@fortawesome/free-solid-svg-icons";
+import { LuChevronLeft, LuXCircle } from "react-icons/lu";
 
-import { useDispatch, useSelector } from 'react-redux';
-import { selectFinishState } from '../redux/finishSlice'
-import { setEnd } from '../redux/endSlice'
+import { useAppDispatch, useAppSelector } from "../redux/store";
+import { selectFinishState } from "../redux/finishSlice";
+import { setEnd } from "../redux/endSlice";
 
-import { PageButton } from './Finish'
+import { PageButton } from "./Finish";
 
-import './../i18n/config';
-import { useTranslation } from 'react-i18next';
-import { selectTheme } from "../redux/themeSlice";
+import { useTranslation } from "react-i18next";
+import { useTheme } from "../themes";
 
 /**
  * Shown if the user wishes to abort.
  * Informs the user about aborting and displays abort button.
  */
-const Discard : React.FC<{}> = () => {
+const Discard: React.FC = () => {
 
   const { t } = useTranslation();
 
-  const finishState = useSelector(selectFinishState)
+  const finishState = useAppSelector(selectFinishState);
 
   const cancelStyle = css({
-    display: finishState !== "Discard changes" ? 'none' : 'flex',
-    flexDirection: 'column' as const,
-    alignItems: 'center',
+    display: finishState !== "Discard changes" ? "none" : "flex",
+    flexDirection: "column" as const,
+    alignItems: "center",
     ...(flexGapReplacementStyle(30, false)),
-  })
+  });
 
   return (
     <div css={cancelStyle}>
@@ -42,39 +38,41 @@ const Discard : React.FC<{}> = () => {
         {t("discard.info-text")}
       </span>
       <div css={backOrContinueStyle}>
-        <PageButton pageNumber={0} label={t("various.goBack-button")} iconName={faChevronLeft} />
+        <PageButton pageNumber={0} label={t("various.goBack-button")} Icon={LuChevronLeft} />
         <DiscardButton />
       </div>
     </div>
   );
-}
+};
 
 /**
  * Button that sets the app into an aborted state
  */
-const DiscardButton : React.FC<{}> = () => {
+const DiscardButton: React.FC = () => {
 
   const { t } = useTranslation();
 
   // Initialize redux variables
-  const dispatch = useDispatch()
-  const theme = useSelector(selectTheme);
+  const dispatch = useAppDispatch();
+  const theme = useTheme();
 
   const discard = () => {
-    dispatch(setEnd({hasEnded: true, value: 'discarded'}))
-  }
+    dispatch(setEnd({ hasEnded: true, value: "discarded" }));
+  };
 
   return (
-    <div css={[basicButtonStyle, navigationButtonStyle(theme)]} title={t("discard.confirm-tooltip")}
+    <div css={[basicButtonStyle(theme), navigationButtonStyle(theme)]}
       role="button" tabIndex={0}
-      onClick={ discard }
-      onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => { if (event.key === " " || event.key === "Enter") {
-        discard()
-      }}}>
-      <FontAwesomeIcon  icon={faTimesCircle} size="1x"/>
+      onClick={discard}
+      onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => {
+        if (event.key === " " || event.key === "Enter") {
+          discard();
+        }
+      }}>
+      <LuXCircle />
       <span>{t("discard.confirm-button")}</span>
     </div>
   );
-}
+};
 
 export default Discard;
