@@ -1,50 +1,48 @@
 import React from "react";
 
-import { css } from '@emotion/react'
+import { css } from "@emotion/react";
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faFrown } from "@fortawesome/free-solid-svg-icons";
+import { LuFrown } from "react-icons/lu";
 
-import { useSelector } from 'react-redux';
-import { selectErrorDetails, selectErrorIcon, selectErrorMessage, selectErrorTitle } from '../redux/errorSlice'
+import { useAppSelector } from "../redux/store";
+import { selectErrorDetails, selectErrorIcon, selectErrorMessage, selectErrorTitle } from "../redux/errorSlice";
 import { flexGapReplacementStyle } from "../cssStyles";
 
-import './../i18n/config';
-import { useTranslation } from 'react-i18next';
+import { useTranslation } from "react-i18next";
 
 /**
  * This page is to be displayed when the application has run into a critical error
  * from which it cannot recover.
  */
- const Error : React.FC<{}> = () => {
+const Error: React.FC = () => {
 
   const { t } = useTranslation();
 
   // Init redux variables
-  const errorTitle = useSelector(selectErrorTitle)
-  const errorMessage = useSelector(selectErrorMessage)
-  const errorDetails = useSelector(selectErrorDetails)
-  const errorIcon = useSelector(selectErrorIcon)
+  const errorTitle = useAppSelector(selectErrorTitle);
+  const errorMessage = useAppSelector(selectErrorMessage);
+  const errorDetails = useAppSelector(selectErrorDetails);
+  const ErrorIcon = useAppSelector(selectErrorIcon);
 
   const detailsStyle = css({
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  })
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+  });
 
   const theEndStyle = css({
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
+    height: "100%",
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
     ...(flexGapReplacementStyle(10, false)),
-  })
+  });
 
   return (
     <div css={theEndStyle} >
       <div>{errorTitle ? errorTitle : t("error.generic-message")}</div>
-      <FontAwesomeIcon icon={errorIcon ? errorIcon : faFrown} size="10x" />
+      {ErrorIcon ? <ErrorIcon css={{ fontSize: 80 }} /> : <LuFrown css={{ fontSize: 80 }} />}
       <span>{errorMessage}</span><br />
       {errorDetails &&
         <div css={detailsStyle}>
@@ -54,6 +52,6 @@ import { useTranslation } from 'react-i18next';
       }
     </div>
   );
-}
+};
 
-export default Error
+export default Error;
