@@ -1,58 +1,55 @@
 import React from "react";
 
-import { css } from '@emotion/react'
-import { basicButtonStyle, flexGapReplacementStyle, tileButtonStyle } from '../cssStyles'
+import { css } from "@emotion/react";
+import { basicButtonStyle, flexGapReplacementStyle, tileButtonStyle } from "../cssStyles";
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faSave, faFileExport, faTimesCircle, IconDefinition
-} from "@fortawesome/free-solid-svg-icons";
+import { IconType } from "react-icons";
+import { LuSave, LuDatabase, LuXCircle } from "react-icons/lu";
 
-import { useDispatch, useSelector } from 'react-redux';
-import { setState, setPageNumber, finish } from '../redux/finishSlice'
+import { useAppDispatch } from "../redux/store";
+import { setState, setPageNumber, finish } from "../redux/finishSlice";
 
-import './../i18n/config';
-import { useTranslation } from 'react-i18next';
-import { selectTheme } from "../redux/themeSlice";
+import { useTranslation } from "react-i18next";
+import { useTheme } from "../themes";
 
 /**
  * Displays a menu for selecting what should be done with the current changes
  */
-const FinishMenu : React.FC<{}> = () => {
+const FinishMenu: React.FC = () => {
 
   const finishMenuStyle = css({
-    display: 'flex',
-    flexDirection: 'row' as const,
-    justifyContent: 'space-around',
-    flexWrap: 'wrap',
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "center",
+    flexWrap: "wrap",
     ...(flexGapReplacementStyle(30, false)),
-  })
+  });
 
   return (
     <div css={finishMenuStyle}>
-        <FinishMenuButton iconName={faSave} stateName="Save changes"/>
-        <FinishMenuButton iconName={faFileExport} stateName="Start processing"/>
-        <FinishMenuButton iconName={faTimesCircle} stateName="Discard changes"/>
+      <FinishMenuButton Icon={LuSave} stateName="Save changes" />
+      <FinishMenuButton Icon={LuDatabase} stateName="Start processing" />
+      <FinishMenuButton Icon={LuXCircle} stateName="Discard changes" />
     </div>
   );
-}
+};
 
 /**
  * Buttons for the finish menu
  */
-const FinishMenuButton: React.FC<{iconName: IconDefinition, stateName: finish["value"]}> = ({iconName, stateName}) => {
+const FinishMenuButton: React.FC<{ Icon: IconType, stateName: finish["value"]; }> = ({ Icon, stateName }) => {
 
   const { t } = useTranslation();
-  const theme = useSelector(selectTheme)
-  const dispatch = useDispatch();
+  const theme = useTheme();
+  const dispatch = useAppDispatch();
 
   const finish = () => {
     dispatch(setState(stateName));
-    dispatch(setPageNumber(1))
-  }
+    dispatch(setPageNumber(1));
+  };
 
-  var buttonString;
-  switch(stateName) {
+  let buttonString;
+  switch (stateName) {
     case "Save changes":
       buttonString = t("finishMenu.save-button");
       break;
@@ -67,15 +64,35 @@ const FinishMenuButton: React.FC<{iconName: IconDefinition, stateName: finish["v
       break;
   }
 
+  const iconStyle = css({
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+
+    background: `${theme.background_finish_menu_icon}`,
+    color: `${theme.text}`,
+    borderRadius: "50%",
+    width: "90px",
+    height: "90px",
+  });
+
+  const labelStyle = css({
+    padding: "0px 20px",
+  });
+
   return (
-    <div css={[basicButtonStyle, tileButtonStyle(theme)]}
-    role="button" tabIndex={0}
-      onClick={ finish }
-      onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => { if (event.key === " " || event.key === "Enter") {
-        finish()
-      }}}>
-      <FontAwesomeIcon  icon={iconName} size="2x"/>
-      <div style={{padding: '0px 20px'}}>{buttonString}</div>
+    <div css={[basicButtonStyle(theme), tileButtonStyle(theme)]}
+      role="button" tabIndex={0}
+      onClick={finish}
+      onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => {
+        if (event.key === " " || event.key === "Enter") {
+          finish();
+        }
+      }}>
+      <div css={iconStyle}>
+        <Icon css={{ fontSize: 36 }} />
+      </div>
+      <div css={labelStyle}>{buttonString}</div>
     </div>
   );
 };
