@@ -1,48 +1,47 @@
 import React from "react";
 
-import { css } from '@emotion/react'
+import { css } from "@emotion/react";
 
-import './../i18n/config';
-import { useTranslation } from 'react-i18next';
+import { useTranslation } from "react-i18next";
 
 /**
  * This page is to be displayed when the application has run into a critical error
  * from which it cannot recover.
  */
- const Landing : React.FC<{}> = () => {
+const Landing: React.FC = () => {
 
   const { t } = useTranslation();
 
   const landingStyle = css({
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
+    height: "100%",
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
 
     a: {
-      color: '#007bff',
-      textDecoration: 'none',
+      color: "#007bff",
+      textDecoration: "none",
     },
     li: {
-      margin: '5px',
+      margin: "5px",
     },
     code: {
-      userSelect: 'all',
-      color: '#e83e8c',
-    }
-  })
+      userSelect: "all",
+      color: "#e83e8c",
+    },
+  });
 
   return (
     <div css={landingStyle} >
       <h1>{t("landing.main-heading")}</h1>
       <div>
         <li>
-        {t("landing.contact-admin")}
+          {t("landing.contact-admin")}
         </li>
         <li>
           {t("landing.start-editing-1")}
-          <code> ?=id=[media-package-id]</code>
+          <code> ?id=[media-package-id]</code>
           {t("landing.start-editing-2")}
         </li>
         <li>
@@ -54,6 +53,6 @@ import { useTranslation } from 'react-i18next';
       </div>
     </div>
   );
-}
+};
 
-export default Landing
+export default Landing;
