@@ -1,203 +1,149 @@
 import React from "react";
 
-import Video from './Video';
-import Timeline from './Timeline';
-import CuttingActions from './CuttingActions';
-import Metadata from './Metadata';
-import TrackSelection from './TrackSelection';
+import Metadata from "./Metadata";
+import TrackSelection from "./TrackSelection";
 import Subtitle from "./Subtitle";
-import Finish from "./Finish"
+import Finish from "./Finish";
 import KeyboardControls from "./KeyboardControls";
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faTools} from "@fortawesome/free-solid-svg-icons";
+import { LuWrench } from "react-icons/lu";
 
-import { css } from '@emotion/react'
+import { css } from "@emotion/react";
 
-import { useSelector } from 'react-redux'
-import { selectMainMenuState } from '../redux/mainMenuSlice'
+import { useAppSelector } from "../redux/store";
+import { selectMainMenuState } from "../redux/mainMenuSlice";
 
-import { MainMenuStateNames } from '../types'
+import { MainMenuStateNames } from "../types";
 import { flexGapReplacementStyle } from "../cssStyles";
 
-import { useBeforeunload } from 'react-beforeunload';
+import { useBeforeunload } from "react-beforeunload";
 import { selectHasChanges as videoSelectHasChanges } from "../redux/videoSlice";
-import { selectHasChanges as metadataSelectHasChanges} from "../redux/metadataSlice";
-import {
-  selectIsPlaying, selectCurrentlyAt,
-  setIsPlaying, setCurrentlyAt, setClickTriggered,
-} from '../redux/videoSlice'
-import { selectTheme } from "../redux/themeSlice";
-import ThemeSwitcher from "./ThemeSwitcher";
+import { selectHasChanges as metadataSelectHasChanges } from "../redux/metadataSlice";
+import { selectHasChanges as selectSubtitleHasChanges } from "../redux/subtitleSlice";
+import { useTheme } from "../themes";
 import Thumbnail from "./Thumbnail";
+import Cutting from "./Cutting";
 
 /**
  * A container for the main functionality
  * Shows different components depending on the state off the app
  */
-const MainContent: React.FC<{}> = () => {
+const MainContent: React.FC = () => {
 
-  const mainMenuState = useSelector(selectMainMenuState)
-  const videoChanged = useSelector(videoSelectHasChanges)
-  const metadataChanged = useSelector(metadataSelectHasChanges)
-  const theme = useSelector(selectTheme)
+  const mainMenuState = useAppSelector(selectMainMenuState);
+  const videoChanged = useAppSelector(videoSelectHasChanges);
+  const metadataChanged = useAppSelector(metadataSelectHasChanges);
+  const subtitleChanged = useAppSelector(selectSubtitleHasChanges);
+  const theme = useTheme();
 
   // Display warning when leaving the page if there are unsaved changes
   useBeforeunload((event: { preventDefault: () => void; }) => {
-    if (videoChanged || metadataChanged) {
+    if (videoChanged || metadataChanged || subtitleChanged) {
       event.preventDefault();
     }
   });
 
-  const cuttingStyle = css({
-    display: 'flex',
-    flexDirection: 'column' as const,
-    justifyContent: 'space-around',
+  const mainContentStyle = css({
+    display: "flex",
+    width: "100%",
+    paddingRight: "20px",
+    paddingLeft: "20px",
     ...(flexGapReplacementStyle(20, false)),
-    paddingRight: '20px',
-    paddingLeft: '161px',
     background: `${theme.background}`,
-  })
+    overflow: "auto",
+  });
+
+  const cuttingStyle = css({
+    flexDirection: "column",
+  });
 
   const metadataStyle = css({
-    display: 'flex',
-    // flexDirection: 'column' as const,
-    // justifyContent: 'space-around',
-    ...(flexGapReplacementStyle(20, false)),
-    paddingRight: '20px',
-    paddingLeft: '161px',
-    background: `${theme.background}`,
-  })
+  });
 
   const trackSelectStyle = css({
-    display: 'flex',
-    flexDirection: 'column' as const,
-    alignContent: 'space-around',
-    ...(flexGapReplacementStyle(20, false)),
-    paddingRight: '20px',
-    paddingLeft: '161px',
-    background: `${theme.background}`,
-  })
+    flexDirection: "column",
+    alignContent: "space-around",
+  });
 
   const subtitleSelectStyle = css({
-    display: 'flex',
-    flexDirection: 'column' as const,
-    justifyContent: 'space-around',
-    paddingRight: '20px',
-    paddingLeft: '161px',
-    height: '100%',
-  })
+    flexDirection: "column",
+    justifyContent: "space-around",
+  });
 
   const thumbnailSelectStyle = css({
-    display: 'flex',
-    flexDirection: 'column' as const,
-    alignContent: 'space-around',
-    ...(flexGapReplacementStyle(20, false)),
-    paddingRight: '20px',
-    paddingLeft: '161px',
-    background: `${theme.background}`,
-  })
+    flexDirection: "column",
+    alignContent: "space-around",
+  });
 
   const finishStyle = css({
-    display: 'flex',
-    flexDirection: 'column' as const,
-    justifyContent: 'space-around',
-    ...(flexGapReplacementStyle(20, false)),
-    paddingRight: '20px',
-    paddingLeft: '161px',
-    minHeight: '100vh',
-    background: `${theme.background}`,
-  })
+    flexDirection: "column",
+    justifyContent: "space-around",
+  });
 
   const keyboardControlsStyle = css({
-    display: 'flex',
-    // flexDirection: 'column' as const,
-    // justifyContent: 'space-around',
-    ...(flexGapReplacementStyle(20, false)),
-    paddingRight: '20px',
-    paddingLeft: '161px',
-    background: `${theme.background}`,
-  })
+    flexDirection: "column",
+  });
 
   const defaultStyle = css({
-    display: 'flex',
-    flexDirection: 'column' as const,
-    alignItems: 'center',
-    padding: '20px',
-    ...(flexGapReplacementStyle(20, false)),
-  })
+    flexDirection: "column",
+    alignItems: "center",
+    padding: "20px",
+  });
 
   const render = () => {
     if (mainMenuState === MainMenuStateNames.cutting) {
       return (
-        <div css={cuttingStyle}>
-          <Video />
-          <CuttingActions />
-          <CuttingTimeline />
+        <div css={[mainContentStyle, cuttingStyle]} role="main">
+          <Cutting />
         </div>
-      )
+      );
     } else if (mainMenuState === MainMenuStateNames.metadata) {
       return (
-        <div css={metadataStyle}>
+        <div css={[mainContentStyle, metadataStyle]} role="main">
           <Metadata />
         </div>
-      )
+      );
     } else if (mainMenuState === MainMenuStateNames.trackSelection) {
       return (
-        <div css={trackSelectStyle}>
+        <div css={[mainContentStyle, trackSelectStyle]} role="main">
           <TrackSelection />
         </div>
-      )
+      );
     } else if (mainMenuState === MainMenuStateNames.subtitles) {
       return (
-        <div css={subtitleSelectStyle}>
+        <div css={[mainContentStyle, subtitleSelectStyle]} role="main">
           <Subtitle />
         </div>
-      )
+      );
     } else if (mainMenuState === MainMenuStateNames.thumbnail) {
       return (
-        <div css={thumbnailSelectStyle}>
+        <div css={[mainContentStyle, thumbnailSelectStyle]} role="main">
           <Thumbnail />
         </div>
-      )
+      );
     } else if (mainMenuState === MainMenuStateNames.finish) {
       return (
-        <div css={finishStyle}>
+        <div css={[mainContentStyle, finishStyle]} role="main">
           <Finish />
         </div>
-        )
+      );
     } else if (mainMenuState === MainMenuStateNames.keyboardControls) {
       return (
-        <div css={keyboardControlsStyle}>
-          <ThemeSwitcher/>
+        <div css={[mainContentStyle, keyboardControlsStyle]} role="main">
           <KeyboardControls />
         </div>
-        )
+      );
     } else {
-      <div css={defaultStyle}>
-        <FontAwesomeIcon icon={faTools} size="10x" />
+      <div css={[mainContentStyle, defaultStyle]} role="main">
+        <LuWrench css={{ fontSize: 80 }} />
         Placeholder
-      </div>
+      </div>;
     }
-  }
+  };
 
   return (
-     <main css={{width: '100%'}} role="main">
-      {render()}
-     </main>
+    <>{render()}</>
   );
 };
-
-const CuttingTimeline : React.FC<{}> = () => {
-  return (
-    <Timeline
-      selectIsPlaying={selectIsPlaying}
-      selectCurrentlyAt={selectCurrentlyAt}
-      setIsPlaying={setIsPlaying}
-      setCurrentlyAt={setCurrentlyAt}
-      setClickTriggered={setClickTriggered}
-    />
-  );
-}
 
 export default MainContent;
