@@ -1,33 +1,25 @@
 import React from "react";
 import SubtitleEditor from "./SubtitleEditor";
 import SubtitleSelect from "./SubtitleSelect";
-import { useSelector } from "react-redux";
+import { useAppSelector } from "../redux/store";
 import { selectIsDisplayEditView } from "../redux/subtitleSlice";
 
 /**
  * A container for the various subtitle views
  */
-const Subtitle : React.FC<{}> = () => {
+const Subtitle: React.FC = () => {
 
-  const displayEditView = useSelector(selectIsDisplayEditView)
+  const displayEditView = useAppSelector(selectIsDisplayEditView);
 
   const render = () => {
-    if (!displayEditView) {
-      return (
-        <SubtitleSelect />
-      )
-    } else {
-      return (
-        <SubtitleEditor />
-      )
-    }
-  }
+    return displayEditView ? <SubtitleEditor /> : <SubtitleSelect />;
+  };
 
   return (
     <>
       {render()}
     </>
   );
-}
+};
 
 export default Subtitle;
