@@ -1,126 +1,119 @@
 import React, { useEffect, useState } from "react";
 import { css } from "@emotion/react";
 import { basicButtonStyle, flexGapReplacementStyle } from "../cssStyles";
-import { faChevronLeft } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { LuChevronLeft, LuDownload } from "react-icons/lu";
 import {
-  selectCaptionTrackByFlavor,
-} from '../redux/videoSlice'
-import { useDispatch, useSelector } from "react-redux";
-import { SubtitleCue } from "../types";
+  selectSubtitlesFromOpencastById,
+} from "../redux/videoSlice";
+import { useAppDispatch, useAppSelector } from "../redux/store";
 import SubtitleListEditor from "./SubtitleListEditor";
 import {
   setIsDisplayEditView,
-  selectSelectedSubtitleByFlavor,
-  selectSelectedSubtitleFlavor,
-  setSubtitle
-} from '../redux/subtitleSlice'
-import { settings } from "../config";
+  selectSelectedSubtitleById,
+  selectSelectedSubtitleId,
+  setSubtitle,
+} from "../redux/subtitleSlice";
 import SubtitleVideoArea from "./SubtitleVideoArea";
 import SubtitleTimeline from "./SubtitleTimeline";
 import { useTranslation } from "react-i18next";
-import { selectTheme } from "../redux/themeSlice";
-import { parseSubtitle } from "../util/utilityFunctions";
+import { useTheme } from "../themes";
+import { parseSubtitle, serializeSubtitle } from "../util/utilityFunctions";
+import { ThemedTooltip } from "./Tooltip";
+import { titleStyle, titleStyleBold } from "../cssStyles";
+import { generateButtonTitle } from "./SubtitleSelect";
 
 /**
  * Displays an editor view for a selected subtitle file
  */
- const SubtitleEditor : React.FC<{}> = () => {
+const SubtitleEditor: React.FC = () => {
 
   const { t } = useTranslation();
 
-  const dispatch = useDispatch()
-  const [getError, setGetError] = useState<string | undefined>(undefined)
-  const subtitle : SubtitleCue[] = useSelector(selectSelectedSubtitleByFlavor)
-  const selectedFlavor = useSelector(selectSelectedSubtitleFlavor)
-  const captionTrack = useSelector(selectCaptionTrackByFlavor(selectedFlavor))
+  const dispatch = useAppDispatch();
+  const [getError, setGetError] = useState<string | undefined>(undefined);
+  const subtitle = useAppSelector(selectSelectedSubtitleById);
+  const selectedId = useAppSelector(selectSelectedSubtitleId);
+  const captionTrack = useAppSelector(selectSubtitlesFromOpencastById(selectedId));
+  const theme = useTheme();
 
   // Prepare subtitle in redux
   useEffect(() => {
     // Parse subtitle data from Opencast
-    if (subtitle === undefined && captionTrack !== undefined && captionTrack.subtitle !== undefined && selectedFlavor) {
+    if (subtitle?.cues === undefined && captionTrack !== undefined && captionTrack.subtitle !== undefined
+      && selectedId) {
       try {
-        dispatch(setSubtitle({identifier: selectedFlavor, subtitles: parseSubtitle(captionTrack.subtitle)}))
+        dispatch(setSubtitle({
+          identifier: selectedId,
+          subtitles: { cues: parseSubtitle(captionTrack.subtitle), tags: captionTrack.tags },
+        }));
       } catch (error) {
         if (error instanceof Error) {
-          setGetError(error.message)
+          setGetError(error.message);
         } else {
-          setGetError(String(error))
+          setGetError(String(error));
         }
       }
 
-    // Or create a new subtitle instead
-    } else if (subtitle === undefined && captionTrack === undefined && selectedFlavor) {
+      // Or create a new subtitle instead
+    } else if (subtitle?.cues === undefined && captionTrack === undefined && selectedId) {
       // Create an empty subtitle
-      dispatch(setSubtitle({identifier: selectedFlavor, subtitles: []}))
+      dispatch(setSubtitle({ identifier: selectedId, subtitles: { cues: [], tags: [] } }));
     }
-  }, [dispatch, captionTrack, subtitle, selectedFlavor])
+  }, [dispatch, captionTrack, subtitle, selectedId]);
 
   const getTitle = () => {
-    return (settings.subtitles.languages !== undefined && subtitle && selectedFlavor) ?
-      settings.subtitles.languages[selectedFlavor] : t("subtitles.editTitle-loading")
-  }
+    if (subtitle) {
+      return generateButtonTitle(subtitle.tags, t);
+    } else {
+      return t("subtitles.editTitle-loading");
+    }
+  };
 
   const subtitleEditorStyle = css({
-    display: 'flex',
-    flexDirection: 'column',
-    paddingRight: '20px',
-    paddingLeft: '20px',
-    gap: '20px',
-    height: '100%',
-  })
+    display: "flex",
+    flexDirection: "column",
+    paddingRight: "20px",
+    paddingLeft: "20px",
+    gap: "20px",
+    height: "100%",
+  });
 
   const headerRowStyle = css({
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    width: '100%',
-  })
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+    width: "100%",
+  });
 
   const subAreaStyle = css({
-    display: 'flex',
-    flexDirection: 'row',
+    display: "flex",
+    flexDirection: "row",
     flexGrow: 1,  // No fixed height, fill available space
-    justifyContent: 'space-between',
-    alignItems: 'top',
-    width: '100%',
-    paddingTop: '10px',
-    paddingBottom: '10px',
+    justifyContent: "space-between",
+    alignItems: "top",
+    width: "100%",
+    paddingTop: "10px",
+    paddingBottom: "10px",
     ...(flexGapReplacementStyle(30, true)),
-    borderBottom: '1px solid #BBB',
-  })
+    borderBottom: `${theme.menuBorder}`,
+  });
 
-  // Taken from VideoHeader. Maybe generalize this to cssStyles.tsx
-  const titleStyle = css({
-    display: 'inline-block',
-    padding: '15px',
-    overflow: 'hidden',
-    whiteSpace: "nowrap",
-    textOverflow: 'ellipsis',
-    maxWidth: '500px',
-  })
-
-  const titleStyleBold = css({
-    fontWeight: 'bold',
-    fontSize: '24px',
-    verticalAlign: '-2.5px',
-  })
 
   const render = () => {
     if (getError !== undefined) {
       return (
         <span>{"Subtitle Parsing Error(s): " + getError}</span>
-      )
+      );
     } else {
-      return(
+      return (
         <>
           <div css={headerRowStyle}>
             <BackButton />
-            <div css={[titleStyle, titleStyleBold]}>
-              {t("subtitles.editTitle", {title: getTitle()})}
+            <div css={[titleStyle(theme), titleStyleBold(theme)]}>
+              {t("subtitles.editTitle", { title: getTitle() })}
             </div>
-            <div css={{width: '50px'}}></div>
+            <DownloadButton />
           </div>
           <div css={subAreaStyle}>
             <SubtitleListEditor />
@@ -128,49 +121,92 @@ import { parseSubtitle } from "../util/utilityFunctions";
           </div>
           <SubtitleTimeline />
         </>
-      )
+      );
     }
-  }
+  };
 
   return (
     <div css={subtitleEditorStyle}>
       {render()}
     </div>
   );
-}
+};
+
+const DownloadButton: React.FC = () => {
+
+  const subtitle = useAppSelector(selectSelectedSubtitleById);
+
+  const downloadSubtitles = () => {
+
+    const vttFile = new Blob([serializeSubtitle(subtitle.cues)], { type: "text/vtt" });
+
+    const vttFileLink = window.URL.createObjectURL(vttFile);
+    const vttHyperLink = document.createElement("a");
+    vttHyperLink.setAttribute("href", vttFileLink);
+
+    const vttFileName = generateButtonTitle(subtitle.tags, t).trimEnd();
+    vttHyperLink.setAttribute("download", `${vttFileName}.vtt`);
+    vttHyperLink.click();
+  };
+
+  const { t } = useTranslation();
+  const theme = useTheme();
+  const style = css({
+    fontSize: "16px",
+    height: "10px",
+    padding: "16px",
+    justifyContent: "space-around",
+    boxShadow: `${theme.boxShadow}`,
+    background: `${theme.element_bg}`,
+  });
+
+  return (
+    <ThemedTooltip title={t("subtitles.downloadButton-tooltip")}>
+      <div css={[basicButtonStyle(theme), style]}
+        role="button"
+        onClick={() => downloadSubtitles()}
+      >
+        <LuDownload css={{ fontSize: "16px" }} />
+        <span>{t("subtitles.downloadButton-title")}</span>
+      </div>
+    </ThemedTooltip>
+  );
+};
 
 
 /**
  * Takes you to a different page
  */
- export const BackButton : React.FC<{}> = () => {
+export const BackButton: React.FC = () => {
 
   const { t } = useTranslation();
-  const theme = useSelector(selectTheme)
-  const dispatch = useDispatch();
+  const theme = useTheme();
+  const dispatch = useAppDispatch();
 
   const backButtonStyle = css({
-    width: '50px',
-    height: '10px',
-    padding: '16px',
+    height: "10px",
+    padding: "16px",
     boxShadow: `${theme.boxShadow}`,
     background: `${theme.element_bg}`,
-    justifyContent: 'space-around'
-  })
+    justifyContent: "space-around",
+  });
 
   return (
-    <div css={[basicButtonStyle, backButtonStyle]}
-      role="button" tabIndex={0}
-      title={t("subtitles.backButton-tooltip")}
-      aria-label={t("subtitles.backButton-tooltip")}
-      onClick={ () => dispatch(setIsDisplayEditView(false)) }
-      onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => { if (event.key === " " || event.key === "Enter") {
-        dispatch(setIsDisplayEditView(false))
-      }}}>
-      <FontAwesomeIcon icon={faChevronLeft} size="1x" />
-      <span>{t("subtitles.backButton")}</span>
-    </div>
+    <ThemedTooltip title={t("subtitles.backButton-tooltip")}>
+      <div css={[basicButtonStyle(theme), backButtonStyle]}
+        role="button" tabIndex={0}
+        aria-label={t("subtitles.backButton-tooltip")}
+        onClick={() => dispatch(setIsDisplayEditView(false))}
+        onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => {
+          if (event.key === " " || event.key === "Enter") {
+            dispatch(setIsDisplayEditView(false));
+          }
+        }}>
+        <LuChevronLeft css={{ fontSize: 24 }} />
+        <span>{t("subtitles.backButton")}</span>
+      </div>
+    </ThemedTooltip>
   );
-}
+};
 
-export default SubtitleEditor
+export default SubtitleEditor;
