@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { css } from "@emotion/react";
-import { useSelector } from "react-redux";
-import { selectCurrentlyAt,
+import { useAppSelector } from "../redux/store";
+import {
+  selectCurrentlyAt,
   selectIsPlaying,
   setClickTriggered,
   setIsPlaying,
@@ -11,22 +12,22 @@ import { selectCurrentlyAt,
   selectAspectRatio,
   setAspectRatio,
   selectCurrentlyAtInSeconds,
-  selectSelectedSubtitleByFlavor,
+  selectSelectedSubtitleById,
   selectIsPlayPreview,
   setIsPlayPreview,
-  setCurrentlyAtAndTriggerPreview} from "../redux/subtitleSlice";
-import { selectVideos } from "../redux/videoSlice";
+  setCurrentlyAtAndTriggerPreview,
+} from "../redux/subtitleSlice";
+import { selectIsMuted, selectVideos, selectVolume, setIsMuted, setVolume } from "../redux/videoSlice";
 import { Flavor } from "../types";
 import { settings } from "../config";
-import { Form } from "react-final-form";
-import { Select } from "mui-rff";
 import { useTranslation } from "react-i18next";
-import { OnChange } from 'react-final-form-listeners'
-import { VideoControls, VideoPlayer } from "./Video";
 import { flexGapReplacementStyle } from "../cssStyles";
 import { serializeSubtitle } from "../util/utilityFunctions";
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-
+import { useTheme } from "../themes";
+import { VideoPlayer } from "./VideoPlayers";
+import VideoControls from "./VideoControls";
+import Select from "react-select";
+import { selectFieldStyle } from "../cssStyles";
 
 /**
  * A part of the subtitle editor that displays a video and related controls
@@ -37,95 +38,99 @@ import { createTheme, ThemeProvider } from '@mui/material/styles';
  * coming up with a proper fix appears to be rather difficult
  * TODO: Come up with a proper fix and create a PR
  */
-const SubtitleVideoArea : React.FC<{}> = () => {
+const SubtitleVideoArea: React.FC = () => {
 
-  const tracks = useSelector(selectVideos)
-  let subtitle = useSelector(selectSelectedSubtitleByFlavor)
-  const [selectedFlavor, setSelectedFlavor] = useState<Flavor>()
-  const [subtitleUrl, setSubtitleUrl] = useState("")
+  const tracks = useAppSelector(selectVideos);
+  const subtitle = useAppSelector(selectSelectedSubtitleById);
+  const [selectedFlavor, setSelectedFlavor] = useState<Flavor>();
+  const [subtitleUrl, setSubtitleUrl] = useState("");
 
   // Decide on initial flavor on mount
   useEffect(() => {
     // Get default from settings
     if (settings.subtitles.defaultVideoFlavor !== undefined) {
-      setSelectedFlavor(settings.subtitles.defaultVideoFlavor)
-      return
+      setSelectedFlavor(settings.subtitles.defaultVideoFlavor);
+      return;
     }
     // If there is no default, just pick any
     if (tracks.length > 0) {
-      setSelectedFlavor(tracks[0].flavor)
-      return
+      setSelectedFlavor(tracks[0].flavor);
+      return;
     }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   // Get the uri of a track the currently selected flavor
   const getTrackURIBySelectedFlavor = () => {
     for (const track of tracks) {
       if (track.flavor.type === selectedFlavor?.type && track.flavor.subtype === selectedFlavor?.subtype) {
-        return track.uri
+        return track.uri;
       }
     }
-  }
+  };
 
   // Get a track URI by any means necessary
   const getTrackURI = () => {
-    const trackURIByFlavor = getTrackURIBySelectedFlavor()
+    const trackURIByFlavor = getTrackURIBySelectedFlavor();
     if (trackURIByFlavor) {
-      return trackURIByFlavor
+      return trackURIByFlavor;
     }
     if (tracks.length > 0) {
-      return tracks[0].uri
+      return tracks[0].uri;
     }
-  }
+  };
 
   // Parse subtitles to something the video player understands
   useEffect(() => {
-    if(subtitle) {
-      const serializedSubtitle = serializeSubtitle(subtitle)
-      setSubtitleUrl(window.URL.createObjectURL(new Blob([serializedSubtitle], {type : 'text/vtt'})))
+    if (subtitle?.cues) {
+      const serializedSubtitle = serializeSubtitle(subtitle?.cues);
+      setSubtitleUrl(window.URL.createObjectURL(new Blob([serializedSubtitle], { type: "text/vtt" })));
     }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [subtitle])
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [subtitle?.cues]);
 
   const areaWrapper = css({
-    display: 'block',
-    height: '100%',
-    width: '40%',
-  })
+    display: "block",
+    height: "100%",
+    width: "40%",
+  });
 
   const videoPlayerAreaStyle = css({
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    height: '100%',
-    // width: '40%',
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
+    height: "100%",
+    // width: "40%",
     ...(flexGapReplacementStyle(10, false)),
   });
 
   const render = () => {
-    return(
+    return (
       <div css={areaWrapper}>
         <div css={videoPlayerAreaStyle}>
-          <VideoSelectDropdown
+          {selectedFlavor && <VideoSelectDropdown
             // eslint-disable-next-line no-sequences
-            flavors={tracks.reduce((a: Flavor[], o: { flavor: Flavor }) => (a.push(o.flavor), a), [])}
+            flavors={tracks.reduce((a: Flavor[], o: { flavor: Flavor; }) => (a.push(o.flavor), a), [])}
             changeFlavorcallback={setSelectedFlavor}
             defaultFlavor={selectedFlavor}
-          />
+          />}
           {/* TODO: Make preview mode work or remove it */}
           <VideoPlayer
             dataKey={0}
             url={getTrackURI()}
             isPrimary={true}
             subtitleUrl={subtitleUrl}
+            first={true}
+            last={true}
             selectIsPlaying={selectIsPlaying}
+            selectIsMuted={selectIsMuted}
             selectCurrentlyAtInSeconds={selectCurrentlyAtInSeconds}
             selectPreviewTriggered={selectPreviewTriggered}
             selectClickTriggered={selectClickTriggered}
             selectAspectRatio={selectAspectRatio}
             setIsPlaying={setIsPlaying}
+            selectVolume={selectVolume}
             setPreviewTriggered={setPreviewTriggered}
             setClickTriggered={setClickTriggered}
             setCurrentlyAt={setCurrentlyAtAndTriggerPreview}
@@ -134,96 +139,90 @@ const SubtitleVideoArea : React.FC<{}> = () => {
           <VideoControls
             selectCurrentlyAt={selectCurrentlyAt}
             selectIsPlaying={selectIsPlaying}
+            selectIsMuted={selectIsMuted}
+            selectVolume={selectVolume}
             selectIsPlayPreview={selectIsPlayPreview}
             setIsPlaying={setIsPlaying}
+            setIsMuted={setIsMuted}
+            setVolume={setVolume}
             setIsPlayPreview={setIsPlayPreview}
           />
         </div>
       </div>
     );
-  }
+  };
 
   return (
     <>
       {render()}
     </>
   );
-}
+};
 
 /**
  * Changes the selectedFlavor in SubtitleVideoArea
  */
-const VideoSelectDropdown : React.FC<{
+const VideoSelectDropdown: React.FC<{
   flavors: Flavor[],
   changeFlavorcallback: React.Dispatch<React.SetStateAction<Flavor | undefined>>,
-  defaultFlavor: Flavor | undefined
+  defaultFlavor: Flavor;
 }> = ({
   flavors,
   changeFlavorcallback,
-  defaultFlavor
+  defaultFlavor,
 }) => {
 
   const { t } = useTranslation();
-  const isDarkPreferred = document.documentElement.getAttribute('data-theme');
+  const theme = useTheme();
 
-  const dropdownName: string = "flavors"
+  const dropdownName = "flavors";
 
   // Turn flavor into string
   const stringifyFlavor = (flavor: Flavor) => {
-    return flavor.type + "/" + flavor.subtype
-  }
+    return flavor.type + "/" + flavor.subtype;
+  };
+
+  const getFlavorLabel = (flavor: Flavor) => {
+    // Omit subtype if all flavour subtypes are equal
+    if (flavors.every(f => f.subtype === flavors[0].subtype)) {
+      return flavor.type;
+    }
+
+    return stringifyFlavor(flavor);
+  };
 
   // Data to populate the dropdown with
-  const selectData = () => {
-    const data = []
-    for (let flavor of flavors) {
-      // We have to deconstruct the flavor object for the value as well and put it back together
-      data.push({label: stringifyFlavor(flavor), value: stringifyFlavor(flavor)})
-    }
-    return data
-  }
-
-  const onSubmit = () => {}
-
-  const muiTheme = createTheme({
-    palette: {
-      mode: isDarkPreferred === 'dark' ? 'dark' : 'light',
-    },
-  });
+  const data = flavors.map(flavor => ({
+    label: getFlavorLabel(flavor),
+    value: stringifyFlavor(flavor),
+  }));
 
   const subtitleAddFormStyle = css({
-    width: '100%',
+    width: "100%",
   });
 
   return (
-    <Form
-    onSubmit={(onSubmit)}
-    // TODO: Find out why "dropdownName" does not work with initialValues
-    initialValues={{"flavors": defaultFlavor ? stringifyFlavor(defaultFlavor) : ""}}
-    render={({ handleSubmit, form, submitting, pristine, values}) => (
-      <form onSubmit={event => {
-        handleSubmit(event)
-      }} css={subtitleAddFormStyle}>
-
-          <ThemeProvider theme={muiTheme}>
-            <Select
-              label={t("subtitleVideoArea.selectVideoLabel")}
-              name={dropdownName}
-              data={selectData()}
-            />
-          </ThemeProvider>
-
-          <OnChange name={dropdownName}>
-            {(value, previous) => {
+    <>
+      <div>{t("subtitleVideoArea.selectVideoLabel")}</div>
+      <Select
+        name={dropdownName}
+        styles={selectFieldStyle(theme)}
+        css={subtitleAddFormStyle}
+        options={data}
+        defaultValue={data.filter(({ value }) => value === stringifyFlavor(defaultFlavor))}
+        onChange={
+          newValue => {
+            if (newValue) {
               // Put flavor back together
-              const newFlavor: Flavor = {type: value.split("/")[0], subtype: value.split("/")[1]}
-              changeFlavorcallback(newFlavor)
-            }}
-          </OnChange>
-      </form>
-    )}
-  />
-  )
-}
+              const [type, subtype] = newValue.value.split("/");
+              const newFlavor: Flavor = { type, subtype };
+              changeFlavorcallback(newFlavor);
+            }
+          }
+        }
+      />
+    </>
+  );
+};
 
 export default SubtitleVideoArea;
