@@ -1,6 +1,6 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit'
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
-import { MainMenuStateNames} from '../types'
+import { MainMenuStateNames } from "../types";
 
 export interface mainMenu {
   value: MainMenuStateNames,
@@ -8,26 +8,27 @@ export interface mainMenu {
 
 const initialState: mainMenu = {
   value: MainMenuStateNames.cutting,
-}
+};
 
 /**
  * Slice for the main menu state
  */
 export const mainMenuSlice = createSlice({
-  name: 'mainMenuState',
+  name: "mainMenuState",
   initialState,
   reducers: {
     setState: (state, action: PayloadAction<mainMenu["value"]>) => {
       state.value = action.payload;
-    }
-  }
-})
+    },
+  },
+});
 
-export const { setState, } = mainMenuSlice.actions
+export const { setState } = mainMenuSlice.actions;
 
 // The function below is called a selector and allows us to select a value from
-// the state. Selectors can also be defined inline where they're used instead of
+// the state. Selectors can also be defined inline where they"re used instead of
 // in the slice file. For example: `useSelector((state) => state.counter.value)`
-export const selectMainMenuState = (state: { mainMenuState: { value: mainMenu["value"]; }; }) => state.mainMenuState.value
+export const selectMainMenuState = (state: { mainMenuState: { value: mainMenu["value"]; }; }) =>
+  state.mainMenuState.value;
 
-export default mainMenuSlice.reducer
+export default mainMenuSlice.reducer;
