@@ -1,14 +1,14 @@
-import { configureStore } from '@reduxjs/toolkit'
-import mainMenuStateReducer from './mainMenuSlice'
-import finishStateReducer from './finishSlice'
-import videoReducer from './videoSlice'
-import workflowPostReducer from './workflowPostSlice'
-import workflowPostAndProcessReducer from './workflowPostAndProcessSlice'
-import endReducer from './endSlice'
-import metadataReducer from './metadataSlice'
-import subtitleReducer from './subtitleSlice'
-import errorReducer from './errorSlice'
-import themeReducer from './themeSlice'
+import { configureStore } from "@reduxjs/toolkit";
+import mainMenuStateReducer from "./mainMenuSlice";
+import finishStateReducer from "./finishSlice";
+import videoReducer from "./videoSlice";
+import workflowPostReducer from "./workflowPostSlice";
+import workflowPostAndProcessReducer from "./workflowPostAndProcessSlice";
+import endReducer from "./endSlice";
+import metadataReducer from "./metadataSlice";
+import subtitleReducer from "./subtitleSlice";
+import errorReducer from "./errorSlice";
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 
 export const store = configureStore({
   reducer: {
@@ -21,11 +21,16 @@ export const store = configureStore({
     metadataState: metadataReducer,
     subtitleState: subtitleReducer,
     errorState: errorReducer,
-    themeState: themeReducer,
-  }
-})
+  },
+});
 
-export default store;
+export type AppDispatch = typeof store.dispatch;
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
-export type RootState = ReturnType<typeof store.getState>
+export type RootState = ReturnType<typeof store.getState>;
+
+// Use instead of plain `useDispatch` and `useSelector`
+export const useAppDispatch: () => AppDispatch = useDispatch;
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+
+export default store;
