@@ -1,53 +1,58 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit'
-import { client } from '../util/client'
-import { Segment, PostEditArgument, httpRequestState } from '../types'
-import { settings } from '../config'
+import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
+import { client } from "../util/client";
+import { Segment, PostEditArgument, httpRequestState } from "../types";
+import { settings } from "../config";
 
 const initialState: httpRequestState = {
-  status: 'idle',
+  status: "idle",
   error: undefined,
-  errorReason: 'unknown',
-}
+  errorReason: "unknown",
+};
 
-export const postVideoInformation = createAsyncThunk('video/postVideoInformation', async (argument: PostEditArgument) => {
-  if (!settings.id) {
-    throw new Error("Missing media package id")
-  }
+export const postVideoInformation =
+  createAsyncThunk("video/postVideoInformation", async (argument: PostEditArgument) => {
+    if (!settings.id) {
+      throw new Error("Missing media package id");
+    }
 
-  const response = await client.post(`${settings.opencast.url}/editor/${settings.id}/edit.json`,
-    { segments: convertSegments(argument.segments), tracks: argument.tracks, subtitles: argument.subtitles }
-  )
-  return response
-})
+    const response = await client.post(`${settings.opencast.url}/editor/${settings.id}/edit.json`,
+      {
+        segments: convertSegments(argument.segments),
+        tracks: argument.tracks,
+        subtitles: argument.subtitles,
+      }
+    );
+    return response;
+  });
 
 /**
  * Slice for managing a post request for saving current changes
  * TODO: Create a wrapper for this and workflowPostAndProcessSlice
  */
 const workflowPostSlice = createSlice({
-  name: 'workflowPostState',
+  name: "workflowPostState",
   initialState,
   reducers: {
-    resetPostRequestState: (state) => {
-      state.status = 'idle'
-    }
+    resetPostRequestState: state => {
+      state.status = "idle";
+    },
   },
   extraReducers: builder => {
     builder.addCase(
-      postVideoInformation.pending, (state, action) => {
-        state.status = 'loading'
-    })
+      postVideoInformation.pending, (state, _action) => {
+        state.status = "loading";
+      });
     builder.addCase(
-      postVideoInformation.fulfilled, (state, action) => {
-        state.status = 'success'
-    })
+      postVideoInformation.fulfilled, (state, _action) => {
+        state.status = "success";
+      });
     builder.addCase(
       postVideoInformation.rejected, (state, action) => {
-        state.status = 'failed'
-        state.error = action.error.message
-    })
-  }
-})
+        state.status = "failed";
+        state.error = action.error.message;
+      });
+  },
+});
 
 interface segmentAPI {
   start: number,
@@ -59,7 +64,7 @@ interface segmentAPI {
 // Convert a segment from how it is stored in redux into
 // a segment that can be send to Opencast
 export const convertSegments = (segments: Segment[]) => {
-  let newSegments: segmentAPI[] = []
+  const newSegments: segmentAPI[] = [];
 
   segments.forEach(segment => {
     newSegments.push({
@@ -67,17 +72,17 @@ export const convertSegments = (segments: Segment[]) => {
       end: segment.end,
       deleted: segment.deleted,
       selected: false,
-    })
+    });
   });
 
-  return newSegments
-}
+  return newSegments;
+};
 
-export const { resetPostRequestState } = workflowPostSlice.actions
+export const { resetPostRequestState } = workflowPostSlice.actions;
 
-export const selectStatus = (state: { workflowPostState: { status: httpRequestState["status"] } }) =>
-  state.workflowPostState.status
-export const selectError = (state: { workflowPostAndProcessState: { error: httpRequestState["error"] } }) =>
-  state.workflowPostAndProcessState.error
+export const selectStatus = (state: { workflowPostState: { status: httpRequestState["status"]; }; }) =>
+  state.workflowPostState.status;
+export const selectError = (state: { workflowPostAndProcessState: { error: httpRequestState["error"]; }; }) =>
+  state.workflowPostAndProcessState.error;
 
-export default workflowPostSlice.reducer
+export default workflowPostSlice.reducer;
