@@ -39,16 +39,24 @@ export interface TimelineState {
 }
 
 export interface SubtitlesFromOpencast {
-  flavor: Flavor,
+  id: string,
   subtitle: string,
+  tags: string[],
+}
+
+export interface SubtitlesInEditor {
+  cues: SubtitleCue[],
+  tags: string[],
 }
 
 export interface SubtitleCue {
-  id: string,
+  id?: string,              // Actually not useful as an identifier, as it is not guaranteed to exist
+  idInternal: string,       // Identifier for internal use. Has nothing to do with the webvtt parser.
   text: string,
   startTime: number,
   endTime: number,
-  tree: {children: [{type: string, value: string}]}   // Odditiy of the webvtt parser. Changes to text also need to be applied to tree.children[0].value
+  // Odditiy of the webvtt parser. Changes to text also need to be applied to tree.children[0].value
+  tree: {children: [{type: string, value: string}]}
   // And many more
 }
 
@@ -65,7 +73,7 @@ export interface ExtendedSubtitleCue extends SubtitleCue {
 export interface PostEditArgument {
   segments: Segment[]
   tracks: Track[]
-  subtitles: {flavor: Flavor, subtitle: string}[]
+  subtitles: SubtitlesFromOpencast[]
 }
 
 export interface PostAndProcessEditArgument extends PostEditArgument{
@@ -84,7 +92,7 @@ export enum MainMenuStateNames {
 }
 
 export interface httpRequestState {
-  status: 'idle' | 'loading' | 'success' | 'failed',
+  status: "idle" | "loading" | "success" | "failed",
   error: string | undefined,
-  errorReason: 'unknown' | 'workflowActive'
+  errorReason: "unknown" | "workflowActive"
 }
